requirements.txt
python-dotenv
langchain-openai
langchain-core
pydantic
langgraph

# streamlit
# datasets
# pandas
# numpy